requirements.txt
streamlit==1.22.0
tensorflow-cpu==2.12.0
numpy==1.23.5
opencv-python-headless==4.6.0.66
Pillow==9.3.0
protobuf==3.20.3  
gdown==4.6.0 
